pyproject.toml
[project]
name = "flask-notes-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Yang Di",email = "yang.di2015@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "marshmallow (>=4.0.0,<5.0.0)",
    "flask-smorest (>=0.46.1,<0.47.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
]

[tool.poetry]
packages = [{include = "flask_notes_app", from = "src"}]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.1.0"
flake8 = "^7.3.0"

[tool.black]
line-length = 79
exclude = '''
/(
    \.venv
  | build
  | dist
  | __pycache__
  | migrations
)/
'''


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
